correct mistakes after mentor check

+++ src/services/contacts.js
@@ -10,11 +10,7 @@ export const getContacts = async ({
   filters = {},
 }) => {
   const skip = (page - 1) * perPage;
-  const contactQuery = ContactCollection.find();
-
-  if (filters.userId) {
-    contactQuery.where('userId').equals(filters.userId);
-  }
+  const contactQuery = ContactCollection.find({ userId: filters.userId });
 
   if (typeof filters.isFavourite === 'boolean') {
     contactQuery.where('isFavourite').equals(filters.isFavourite);
@@ -23,14 +19,15 @@ export const getContacts = async ({
     contactQuery.where('contactType').equals(filters.contactType);
   }
 
-  const totalItems = await ContactCollection.find();
-
   const data = await contactQuery
     .skip(skip)
     .limit(perPage)
     .sort({ [sortBy]: sortOrder })
-    .merge(contactQuery)
-    .countDocuments();
+    .merge(contactQuery);
+
+  const totalItems = await ContactCollection.find({
+    userId: filters.userId,
+  }).countDocuments();
 
   const paginationData = calcPaginationData({ page, perPage, totalItems });
 
@@ -43,16 +40,25 @@ export const getContacts = async ({
   };
 };
 
-export const getContactById = (id) => ContactCollection.findOne({ _id: id });
+export const getContactById = (contactId, userId) =>
+  ContactCollection.findOne({ _id: contactId, userId });
 
 export const addContact = (payload) => ContactCollection.create(payload);
 
-export const updateContact = async (_id, payload, options = {}) => {
+export const updateContact = async (
+  { contactId, userId },
+  payload,
+  options = {},
+) => {
   const { upsert = false } = options;
-  const rawResult = await ContactCollection.findOneAndUpdate({ _id }, payload, {
-    upsert,
-    includeResultMetadata: true,
-  });
+  const rawResult = await ContactCollection.findOneAndUpdate(
+    { _id: contactId, userId },
+    payload,
+    {
+      upsert,
+      includeResultMetadata: true,
+    },
+  );
 
   if (!rawResult || !rawResult.value) return null;
 
@@ -62,5 +68,5 @@ export const updateContact = async (_id, payload, options = {}) => {
   };
 };
 
-export const deleteContactById = (_id) =>
-  ContactCollection.findOneAndDelete({ _id });
+export const deleteContactById = (contactId, userId) =>
+  ContactCollection.findOneAndDelete({ _id: contactId, userId });
